refactor(input): ♻️ renamed folder view to views
