+++ src/components/DatePicker/types/DatePicker.ts
@@ -1,5 +1,5 @@
 import { ReactDatePickerProps } from 'react-datepicker';
-import { InputProps } from '../../Input/view/base/types';
+import { InputProps } from '../../Input/views/base/types';
 
 export type DatePickerProps = {} & ReactDatePickerProps &
     Pick<InputProps, 'isFullWidth'>;
