+++ src/components/Input/index.ts
@@ -1,2 +1,2 @@
 export * from './hooks';
-export * from './view';
+export * from './views';
